+++ sematic/db/models/tests/test_resolution.py
@@ -6,10 +6,11 @@ import pytest
 # Sematic
 from sematic.db.models.resolution import (
     InvalidResolution,
-    Resolution,
     ResolutionKind,
     ResolutionStatus,
 )
+from sematic.db.tests.fixtures import make_resolution
+from sematic.utils.git import GitInfo
 
 
 def test_is_allowed_transition():
@@ -26,70 +27,87 @@ def test_is_allowed_transition():
 
 UPDATE_CASES = [
     (
-        Resolution(
+        make_resolution(
             root_id="abc123",
             status=ResolutionStatus.SCHEDULED,
             kind=ResolutionKind.KUBERNETES,
             docker_image_uri="my.docker.registry.io/image/tag",
+            git_info=GitInfo(remote="r", branch="b", commit="c", dirty=False),
         ),
         None,
     ),
     (
-        Resolution(
+        make_resolution(
             root_id="abc123",
             status=ResolutionStatus.RUNNING,
             kind=ResolutionKind.KUBERNETES,
             docker_image_uri="my.docker.registry.io/image/tag",
+            git_info=GitInfo(remote="r", branch="b", commit="c", dirty=False),
         ),
         None,
     ),
     (
-        Resolution(
+        make_resolution(
             root_id="zzz",
             status=ResolutionStatus.SCHEDULED,
             kind=ResolutionKind.KUBERNETES,
             docker_image_uri="my.docker.registry.io/image/tag",
+            git_info=GitInfo(remote="r", branch="b", commit="c", dirty=False),
         ),
         r"Cannot update root_id of resolution abc123 after it has been created.*zzz.*",
     ),
     (
-        Resolution(
+        make_resolution(
             root_id="abc123",
             status=ResolutionStatus.COMPLETE,
             kind=ResolutionKind.KUBERNETES,
             docker_image_uri="my.docker.registry.io/image/tag",
+            git_info=GitInfo(remote="r", branch="b", commit="c", dirty=False),
         ),
         r"Resolution abc123 cannot be moved from the SCHEDULED state to the "
         r"COMPLETE state.",
     ),
     (
-        Resolution(
+        make_resolution(
             root_id="abc123",
             status=ResolutionStatus.SCHEDULED,
             kind=ResolutionKind.LOCAL,
             docker_image_uri="my.docker.registry.io/image/tag",
+            git_info=GitInfo(remote="r", branch="b", commit="c", dirty=False),
         ),
         r"Cannot update kind of resolution abc123 after it has been created.*LOCAL.*",
     ),
     (
-        Resolution(
+        make_resolution(
             root_id="abc123",
             status=ResolutionStatus.SCHEDULED,
             kind=ResolutionKind.KUBERNETES,
             docker_image_uri="my.docker.registry.io/changed/tag",
+            git_info=GitInfo(remote="r", branch="b", commit="c", dirty=False),
         ),
         r"Cannot update docker_image_uri of resolution abc123 .*changed/tag.*",
+    ),
+    (
+        make_resolution(
+            root_id="abc123",
+            status=ResolutionStatus.SCHEDULED,
+            kind=ResolutionKind.KUBERNETES,
+            docker_image_uri="my.docker.registry.io/image/tag",
+            git_info=GitInfo(remote="r", branch="b", commit="c", dirty=True),
+        ),
+        r"Cannot update git_info_json of resolution abc123 .*\"dirty\": false.*",
     ),
 ]
 
 
 @pytest.mark.parametrize("update,expected_error", UPDATE_CASES)
 def test_updates(update, expected_error):
-    original = Resolution(
+    original = make_resolution(
         root_id="abc123",
         status=ResolutionStatus.SCHEDULED,
         kind=ResolutionKind.KUBERNETES,
         docker_image_uri="my.docker.registry.io/image/tag",
+        git_info=GitInfo(remote="r", branch="b", commit="c", dirty=False),
     )
     try:
         original.update_with(update)