+++ docs/changelog.md
@@ -4,6 +4,7 @@ Lines for version numbers should always be formatted as `* MAJOR.MINOR.PATCH`
 with nothing else on the line.
 -->
 * HEAD
+    * [feature] Expose pipeline git commit info in the UI
 * 0.16.0
     * [feature] Support Enums
     * [feature] Allow specifying Kubernetes tolerations for cloud jobs